requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
streamlit
pillow
requests
fal-client
streamlit-cookies-manager